pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ssp_transformations_handler"
version = "0.1.0"
description = "A package to handle SSP model transformations"
authors = [{ name = "Juan Antonio Robledo Lara", email = "jroblar98@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "pandas",
    "pyyaml"
]

[tool.setuptools]
packages = ["ssp_transformations_handler"]

[project.scripts]
ssp_tool = "ssp_transformations_handler.cli:main"  # (Optional) CLI command
